ONE BUG TO FIX AND PROJECT IS DONE

+++ src/com/userfiles/BaseAccount.java
@@ -12,13 +12,13 @@ public class BaseAccount extends Account implements Serializable {
 
     @Override
     public void addToBalance(Transaction transaction) {
-        this.setBalance((this.getBalance()*100+transaction.getAmount()*100)/100);
+        this.setBalance(((int) (this.getBalance()*100+transaction.getAmount()*100))/100.0);
         this.getTransactions().add(transaction);
     }
 
     @Override
     public void subFromBalance(Transaction transaction) {
-        this.setBalance((this.getBalance()*100-transaction.getAmount()*100)/100);
+        this.setBalance(((int) (this.getBalance()*100-transaction.getAmount()*100))/100.0);
         this.getTransactions().add(transaction);
     }
 }
